requirements.txt
flask
python-dotenv
requests
pyodbc
flask_sqlalchemy
azure-storage-blob
msal
azure-identity
pyjwt